some sizeof stuff

+++ test/input.cpp
@@ -1,4 +1,4 @@
-int main(){
+float main(){
     float a = 1.1;
     float b = 2.2f;
     return a+b;
